refactor(component): pilltabs - move useAvailableWidth to own file
